Refactor CSS for login page and nomination show comonent (#147)
* moved css for components into respective style.css file instead of app.css

* deleted index.css and added content to App.css, refactored login page and nomination show page css, moved google btn to login page directory

* deleted old google btn component

* uncommented out body and code tags

* changed banner background, border radius, and margin

* added color to body font and h2, add font body and secondary dark classes to nomiation banner to fix color issues

* fixed col sizing
