+++ packages/app/src/components/nomination/nominationBanner.js
@@ -14,7 +14,7 @@ const NominationBanner = ({ nomination }) => {
           <img src="/ksflogo.png" alt="other"/>
         </div>
 
-        <div className="column column-90 nomination-details">
+        <div className="column column-80 nomination-details">
 
           <div className="row banner-top">
             <div className="column name">
@@ -25,20 +25,20 @@ const NominationBanner = ({ nomination }) => {
 
           <div className="row banner-bottom">
             <div className="column hp-name">
-              <p>HP Name</p>
-              <span><h2><strong>{nomination.providerName}</strong></h2></span>
+              <p className="secondary-dark">HP Name</p>
+              <span><h2 className="body-font"><strong>{nomination.providerName}</strong></h2></span>
             </div>
             <div className="column fam-name">
-              <p>Family Member Name</p>
-              <span><h2><strong>{nomination.representativeName}</strong></h2></span>
+              <p className="secondary-dark">Family Member Name</p>
+              <span><h2 className="body-font"><strong>{nomination.representativeName}</strong></h2></span>
             </div>
             <div className="column created-at">
-              <p>Created Date</p>
-              <span><h2><strong>{date}</strong></h2></span>
+              <p className="secondary-dark">Created Date</p>
+              <span><h2 className="body-font"><strong>{date}</strong></h2></span>
             </div>
             <div className="column amount">
-               <p>Grant Amount Requested</p>
-              <span><h2><strong>${nomination.amountRequestedCents}</strong></h2></span>
+               <p className="secondary-dark">Grant Amount Requested</p>
+              <span><h2 className="body-font"><strong>${nomination.amountRequestedCents}</strong></h2></span>
             </div>
           </div>
         </div>
