+++ packages/app/src/components/pages/Login/GoogleBtn/GoogleBtn.js
@@ -1,6 +1,5 @@
 import React, { Component } from "react";
 import { GoogleLogin, GoogleLogout } from "react-google-login";
-import '../../App.css'
 
 class GoogleBtn extends Component {
   state = {
