+++ packages/app/src/components/pages/Login/style.css
@@ -1,0 +1,17 @@
+@import "../../../App.css";
+
+/*LOGIN GOOGLE BUTTON*/
+.google-button {
+  text-align: center;
+}
+
+.google-login, .google-logout {
+  /* using important to override inline styles rendered by 3rd party component */
+  padding: 30px !important;
+}
+
+.login-logo {
+  max-width: 25%;
+  margin: 30% auto;
+  margin-bottom: 5%;
+}
